MKTAXI-2926 => Code react

+++ Src/Server/MK.Booking.Api.Contract/Resources/AvailableVehicle.cs
@@ -8,10 +8,7 @@
         public string LogoName { get; set; }
         public int FleetId { get; set; }
 
-        /// <summary>
-        /// This value is in seconds
-        /// </summary>
-        public long? Eta { get; set; }
+        public int? Eta { get; set; }
     }
 
     public class AvailableVehicleCluster : AvailableVehicle
